pyproject.toml
[project]
name = "inprompt"
version = "1.2.0"
description = "Simple CLI to output source code blocks for LLM prompts."
authors = [{ name = "Dominik Roblek" }]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "absl-py",
    "loguru",
]
license = "MIT"

keywords = ["prompt", "code", "formatting"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Topic :: Utilities",
    "License :: OSI Approved :: MIT License",
]

[project.scripts]
inprompt = "inprompt:run"

[tool.poetry.group.dev.dependencies]
ruff = "*"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
# Tells Poetry we do NOT have a traditional package folder,
# but DO want to include the single-file module `inprompt.py`.
packages = []
include = ["inprompt.py"]

[tool.ruff]
lint.select = ["E", "F", "I"]
